Update Dockerfile and webpack build to ensure app gets STAGE env.
The React app checks process.env.STAGE to determine whether or not
to use the stage endpoints for data queries and more. Although it's
possible for the React app on the server side (a node process) to
check its environment at run time, this is an expensive check and can
slow the response time. On the browser side, some other check than to
process.env must be invented. (I believe) For these reasons, the
ReactQL webpack config is written such that all references to the
native process.env are replaced with a plain Javascript object.

Therefore, the distinction for whether the image is to be run for
stage or prod must be known at image build time (which is the same
time when the app distribution is built by Webpack). I've updated
the Dockerfile to take in an bool argument STAGE to determine whether
or not to set process.env.STAGE to true or false during Webpack build.