Ensure that requests to server come in with null jwt token.
This results in an anonymous user being hydrated on the backend,
such that the javascript response renders accordingly. Without
setting the jwt token, views regarding user logic cannot be relied
on rendering properly.

In the future, it may be a goal to sync the token stored locally on
the client/user's side with a store in the server, so that the re-
sponse can be rendered for a given user from the start (i.e. support
full user sessions). But for now, hydrating an anonymous user means
that the first view upon hitting moshi-moji.xyz will be for no user
logged in; then once the browser bundle renders, it should feature
the proper user so long as that user has logged in before.