[Bug fix] Temporarily fix fragment creation during build.
This change will ensure that fragments are created against the
stage endpoint for graphql.
Previously the fragment json generation would fail if localhost
was not serving graphql.
However, when/if the schemas ever differ between local, prod and
stage, the fragments will be incorrect. In the future this script
should determine the environment and run against the corresponding
graphql endpoint.